add functionality to persist progress

+++ backend/src/main/kotlin/com/memento/flashcards/FlashCardRepository.kt
@@ -4,7 +4,7 @@ import com.memento.course.Course
 import org.springframework.data.jpa.repository.JpaRepository
 import java.util.*
 
-interface FlashCardRepository: JpaRepository<FlashCard, Long> {
+interface FlashCardRepository: JpaRepository<FlashCard, Int> {
     fun findAllByCourseOrderById(course: Course): List<FlashCard>
     fun deleteByCourseIdAndId(courseId: Int, flashCardId: Int): Void
     fun findByCourseAndId(course: Course, flashCardId: Int): Optional<FlashCard>
